server/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
shap
google-generativeai
python-dotenv
